requirements.txt
tensorflow~=2.5.0
opencv-python~=4.5.2.54
numpy~=1.19.5
sklearn~=0.0
matplotlib~=3.4.2

pandas~=1.3.0
scikit-learn~=0.24.2
imgaug~=0.4.0
python-engineio=3.13.2
eventlet~=0.31.1
Pillow~=8.3.1
Flask~=2.0.1